pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oslo"
version = "0.1.0"
description = "Generate short vertical videos from text using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "openai>=1.30",
    "moviepy>=2.0",
    "pillow>=10.0",
    "pydub>=0.25",
    "audioop-lts>=0.2; python_version>='3.13'",
    "python-dotenv>=1.0",
]

[project.scripts]
oslo = "oslo.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
